+++ src/realm/group_shared.cpp
@@ -50,6 +50,7 @@
 
 
 using namespace realm;
+using namespace realm::metrics;
 using namespace realm::util;
 using Durability = SharedGroupOptions::Durability;
 
@@ -772,6 +773,13 @@ void SharedGroup::do_open(const std::string& path, bool no_create_file, bool is_
     m_lockfile_prefix = m_coordination_dir + "/access_control";
     SlabAlloc& alloc = m_group.m_alloc;
 
+#if REALM_METRICS
+    if (options.enable_metrics) {
+        m_metrics = std::make_shared<Metrics>();
+        m_group.set_metrics(m_metrics);
+    }
+#endif // REALM_METRICS
+
     Replication::HistoryType openers_hist_type = Replication::hist_None;
     int openers_hist_schema_version = 0;
     bool opener_is_sync_agent = false;
@@ -1233,7 +1241,7 @@ void SharedGroup::do_open(const std::string& path, bool no_create_file, bool is_
         break;
     }
 
-    m_transact_stage = transact_Ready;
+    set_transact_stage(transact_Ready);
 // std::cerr << "open completed" << std::endl;
 
 #ifdef REALM_ASYNC_DAEMON
@@ -1376,7 +1384,7 @@ void SharedGroup::close() noexcept
             break;
     }
     m_group.detach();
-    m_transact_stage = transact_Ready;
+    set_transact_stage(transact_Ready);
     SharedInfo* info = m_file_map.get_addr();
     {
         bool is_sync_agent = false;
@@ -1458,6 +1466,35 @@ void SharedGroup::enable_wait_for_change()
 {
     std::lock_guard<InterprocessMutex> lock(m_controlmutex);
     m_wait_for_change_enabled = true;
+}
+
+void SharedGroup::set_transact_stage(SharedGroup::TransactStage stage) noexcept
+{
+#if REALM_METRICS
+    if (m_metrics) { // null if metrics are disabled
+        size_t total_size = m_used_space + m_free_space;
+        size_t free_space = m_free_space;
+        size_t num_objects = m_group.m_total_rows;
+        size_t num_available_versions = static_cast<size_t>(get_number_of_versions());
+
+        if (stage == transact_Reading) {
+            if (m_transact_stage == transact_Writing) {
+                m_metrics->end_write_transaction(total_size, free_space, num_objects, num_available_versions);
+            }
+            m_metrics->start_read_transaction();
+        } else if (stage == transact_Writing) {
+            if (m_transact_stage == transact_Reading) {
+                m_metrics->end_read_transaction(total_size, free_space, num_objects, num_available_versions);
+            }
+            m_metrics->start_write_transaction();
+        } else if (stage == transact_Ready) {
+            m_metrics->end_read_transaction(total_size, free_space, num_objects, num_available_versions);
+            m_metrics->end_write_transaction(total_size, free_space, num_objects, num_available_versions);
+        }
+    }
+#endif
+
+    m_transact_stage = stage;
 }
 
 #ifdef REALM_ASYNC_DAEMON
@@ -1738,7 +1775,7 @@ const Group& SharedGroup::begin_read(VersionID version_id)
     bool writable = false;
     do_begin_read(version_id, writable); // Throws
 
-    m_transact_stage = transact_Reading;
+    set_transact_stage(transact_Reading);
     return m_group;
 }
 
@@ -1755,7 +1792,7 @@ void SharedGroup::end_read() noexcept
 
     do_end_read();
 
-    m_transact_stage = transact_Ready;
+    set_transact_stage(transact_Ready);
 }
 #ifdef _MSC_VER
 #pragma warning (default: 4297)
@@ -1787,7 +1824,7 @@ Group& SharedGroup::begin_write()
         throw;
     }
 
-    m_transact_stage = transact_Writing;
+    set_transact_stage(transact_Writing);
     return m_group;
 }
 
@@ -1817,7 +1854,7 @@ bool SharedGroup::try_begin_write(Group* &group)
         throw;
     }
 
-    m_transact_stage = transact_Writing;
+    set_transact_stage(transact_Writing);
     group = &m_group;
     return true;
 }
@@ -1848,7 +1885,7 @@ SharedGroup::version_type SharedGroup::commit()
 
     do_end_read();
     m_read_lock = lock_after_commit;
-    m_transact_stage = transact_Ready;
+    set_transact_stage(transact_Ready);
     return new_version;
 }
 
@@ -1871,7 +1908,7 @@ void SharedGroup::rollback() noexcept
     if (Replication* repl = m_group.get_replication())
         repl->abort_transact();
 
-    m_transact_stage = transact_Ready;
+    set_transact_stage(transact_Ready);
 }
 #ifdef _MSC_VER
 #pragma warning (default: 4297)
@@ -1898,6 +1935,12 @@ void SharedGroup::unpin_version(VersionID token)
     release_read_lock(read_lock);
 }
 
+#if REALM_METRICS
+std::shared_ptr<Metrics> SharedGroup::get_metrics()
+{
+    return m_metrics;
+}
+#endif // REALM_METRICS
 
 void SharedGroup::do_begin_read(VersionID version_id, bool writable)
 {
@@ -2094,7 +2137,7 @@ SharedGroup::version_type SharedGroup::commit_and_continue_as_read()
     // Remap file if it has grown, and update refs in underlying node structure
     gf::remap_and_update_refs(m_group, m_read_lock.m_top_ref, m_read_lock.m_file_size); // Throws
 
-    m_transact_stage = transact_Reading;
+    set_transact_stage(transact_Reading);
 
     return version;
 }
@@ -2181,6 +2224,10 @@ void SharedGroup::low_level_commit(uint_fast64_t new_version)
     // Do the actual commit
     REALM_ASSERT(m_group.m_top.is_attached());
     REALM_ASSERT(oldest_version <= new_version);
+
+#if REALM_METRICS
+    m_group.update_num_objects();
+#endif // REALM_METRICS
     // info->readers.dump();
     GroupWriter out(m_group); // Throws
     out.set_versions(new_version, oldest_version);